add the ability to define a “key equivalent” on menu items
This is the definition of the short-cut key.
A key equivalent of `","` will be `⌘+,` and `<Ctrl>+,` on Windows or Linux.
Key equivalents currently only work on MacOS native menus.
Currently, they hide definition of short-cuts. So if the key equivalent `"c"` is
defined, the copy short-cut will stop working.
There has to be done more work on integrating this better.