add the ability to place menus before or after additional native menus
On MacOS, there are two native menus: the app menu and the window menu.
The app menu should always be first while the window menu usually is the
last before help.
With this new setting the defined menus can be placed before (default)
or after the native menus (except for app menu which will be always the first).