Add ability to specify custom metadata.
Custom metadata can be defined and picked from the Development or
Release section of FyneApp.toml or specified on the command. In the
later case it will override any custom metadata provided in the
FyneApp.toml. This is useful if you want to provide for example a
remote self update url that depends on if the build is a release or
a development release. Or if you want to provide some kind of API key
during the build and do not want to store it in your repository.