Races fixed up enough to close out the fyne_demo reports
This is not using the final API but I wanted to progress tasks as while working on trickier details.
I will move this to public API when it is decided, a simple refactor will cover it.

In the meantime I can no longer get any reports from  in fyne_demo